[#2374] Moved mixin into utils
